docs: add test instructions
