+++ options.html
@@ -203,6 +203,7 @@
 
   <div id="status" class="status"></div>
 
+  <script src="utils.js"></script>
   <script src="options.js"></script>
 </body>
 </html>
